+++ src/domain/Zone.h
@@ -29,7 +29,7 @@ public:
     bool clearError(uint32_t nowMs);
     bool checkIdleLeak(uint32_t pulseCount, uint32_t nowMs, uint16_t windowSec, uint16_t threshold, uint32_t* observedPulses);
     void resetLeakWindow(uint32_t pulseCount, uint32_t nowMs);
-    Irrigation::ZoneStatus status(uint32_t pulseCount, uint32_t flowRatePerMinuteX1000, uint32_t flowMlPerMin, bool flowRateReady, uint32_t nowMs) const;
+    Irrigation::ZoneStatus status(uint32_t pulseCount, uint64_t flowRatePerMinuteX1000, uint32_t flowMlPerMin, bool flowRateReady, uint32_t nowMs) const;
     const Irrigation::ZoneConfig& config() const;
     Irrigation::ZoneState state() const;
     bool isBusy() const;
